Mac: Fixed several minor bugs in Splash and Preferences.
- SplashWindowController wasn't refreshing the status message.
- PreferencesWindowController wasn't refreshing the Dropbox Login/Logout
button correctly.
- CloudLibraryService wasn't calculating the download progress
correctly.

+++ MPfm/MPfm.Library/Services/CloudLibraryService.cs
@@ -99,7 +99,7 @@ namespace MPfm.Library.Services
                 }
 
                 if (OnDeviceInfosDownloadProgress != null)
-                    OnDeviceInfosDownloadProgress(_deviceInfos.Count/(_deviceInfos.Count + _deviceInfosLeftToDownload.Count));
+                    OnDeviceInfosDownloadProgress((float)_deviceInfos.Count/(float)(_deviceInfos.Count + _deviceInfosLeftToDownload.Count));
 
                 // Check if list is already empty; do not raise OnDeviceInfosAvailable multiple times
 				if (_deviceInfosLeftToDownload.Count == 0)
